Duplicated route controller test removed
